move some of the components into indiviaul class. because seems it's a little bit large in the TranslateEditSection.
